+++ src/components/overview/DescriptionDetails/AddToCart/AddToCart.jsx
@@ -4,7 +4,7 @@ import axios from 'axios';
 import SizeSelector from './SizeSelector';
 import QtySelector from './QtySelector';
 import CartMessage from './CartMessage';
-import config from '../../../config/config';
+import config from '../../../../../config/config';
 
 export default function AddToCart({ skuData, currStyle }) {
   const [currSku, setCurrSku] = useState('Select Size');
@@ -53,7 +53,9 @@ export default function AddToCart({ skuData, currStyle }) {
         .then(() => {
           setMessage('success');
         })
-        .catch((err) => console.error(err));
+        .catch(() => {
+          setMessage('failure');
+        });
     }
   };
