+++ src/components/overview/ImageGallery/ExpandedIcons.jsx
