+++ src/components/overview/ImageGallery/ExpandedView.jsx
