+++ src/components/overview/ImageGallery/Thumbnails.jsx
